# src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "oscp-spatial-service-discovery",
    "version": "1.0.0",
    "description": "OSCP Spatial Service Discovery (Open AR Cloud). Successful record payloads use the vendor media type `application/vnd.oscp+json; version=1.0`.",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Local dev"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check",
        "operationId": "getHealth",
        "tags": ["Health"],
        "responses": {
          "200": {
            "$ref": "#/components/responses/HealthOk"
          }
        }
      }
    },
    "/countries": {
      "get": {
        "summary": "List ISO country codes served by this SSD instance",
        "operationId": "listCountries",
        "tags": ["Health"],
        "responses": {
          "200": {
            "description": "Uppercase ISO country codes from this instance's COUNTRIES configuration",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "type": "string" },
                  "example": ["FI", "IT", "US"]
                }
              }
            }
          }
        }
      }
    },
    "/{country}/ssrs": {
      "get": {
        "summary": "Get SSRs intersecting an H3 index",
        "operationId": "findSsrsByHex",
        "tags": ["SSR"],
        "parameters": [
          { "$ref": "#/components/parameters/CountryPath" },
          { "$ref": "#/components/parameters/H3IndexQueryRequired" }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/OscpSsrArray"
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        }
      },
      "post": {
        "summary": "Create a new SSR",
        "operationId": "createSsr",
        "tags": ["SSR"],
        "parameters": [{ "$ref": "#/components/parameters/CountryPath" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SsrDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created; response body is the new record id",
            "content": {
              "text/plain": {
                "schema": { "type": "string" }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "security": [{ "bearerAuth": ["create:ssrs"] }]
      }
    },
    "/{country}/ssrs/{id}": {
      "get": {
        "summary": "Get SSR by id",
        "operationId": "getSsrById",
        "tags": ["SSR"],
        "parameters": [
          { "$ref": "#/components/parameters/CountryPath" },
          { "$ref": "#/components/parameters/IdPath" }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/OscpSsr"
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        }
      },
      "put": {
        "summary": "Update SSR",
        "operationId": "updateSsr",
        "tags": ["SSR"],
        "parameters": [
          { "$ref": "#/components/parameters/CountryPath" },
          { "$ref": "#/components/parameters/IdPath" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SsrDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated"
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "security": [{ "bearerAuth": ["update:ssrs"] }]
      },
      "delete": {
        "summary": "Delete SSR",
        "operationId": "deleteSsr",
        "tags": ["SSR"],
        "parameters": [
          { "$ref": "#/components/parameters/CountryPath" },
          { "$ref": "#/components/parameters/IdPath" }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "security": [{ "bearerAuth": ["delete:ssrs"] }]
      }
    },
    "/{country}/provider/ssrs": {
      "get": {
        "summary": "List SSRs for the authenticated provider",
        "operationId": "listProviderSsrs",
        "tags": ["SSR"],
        "parameters": [{ "$ref": "#/components/parameters/CountryPath" }],
        "responses": {
          "200": {
            "$ref": "#/components/responses/OscpSsrArray"
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "security": [{ "bearerAuth": ["read:ssrs"] }]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Auth0 JWT. When AUTH_REQUIRED=false (local/dev), protected routes accept requests without a token."
      }
    },
    "parameters": {
      "CountryPath": {
        "name": "country",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "ISO country code (uppercased by the service)"
      },
      "IdPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "Platform-generated SSR id"
      },
      "H3IndexQueryRequired": {
        "name": "h3Index",
        "in": "query",
        "required": true,
        "schema": { "type": "string" },
        "description": "H3 cell index for geographic search"
      }
    },
    "responses": {
      "HealthOk": {
        "description": "Service is healthy",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Health" }
          }
        }
      },
      "OscpSsr": {
        "description": "OK",
        "content": {
          "application/vnd.oscp+json; version=1.0": {
            "schema": { "$ref": "#/components/schemas/Ssr" }
          }
        }
      },
      "OscpSsrArray": {
        "description": "OK",
        "content": {
          "application/vnd.oscp+json; version=1.0": {
            "schema": {
              "type": "array",
              "items": { "$ref": "#/components/schemas/Ssr" }
            }
          }
        }
      },
      "Error400": {
        "description": "Bad request (validation, invalid country, invalid h3Index, invalid polygon, etc.)",
        "content": {
          "text/plain": {
            "schema": { "type": "string" }
          }
        }
      },
      "Error403": {
        "description": "Forbidden (invalid provider / JWT claim)",
        "content": {
          "text/plain": {
            "schema": { "type": "string" }
          }
        }
      },
      "Error404": {
        "description": "Record not found",
        "content": {
          "text/plain": {
            "schema": { "type": "string" }
          }
        }
      },
      "Error500": {
        "description": "Internal server error",
        "content": {
          "text/plain": {
            "schema": { "type": "string" }
          }
        }
      }
    },
    "schemas": {
      "Health": {
        "type": "object",
        "required": ["status"],
        "properties": {
          "status": {
            "type": "string",
            "example": "ok"
          }
        }
      },
      "Property": {
        "type": "object",
        "required": ["type", "value"],
        "properties": {
          "type": { "type": "string" },
          "value": { "type": "string" }
        }
      },
      "Service": {
        "type": "object",
        "required": ["id", "type", "title", "url"],
        "properties": {
          "id": { "type": "string" },
          "type": { "type": "string" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "url": { "type": "string", "format": "uri" },
          "properties": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Property" }
          }
        }
      },
      "Polygon": {
        "type": "object",
        "required": ["type", "coordinates"],
        "properties": {
          "type": {
            "type": "string",
            "enum": ["Polygon"]
          },
          "coordinates": {
            "type": "array",
            "items": {
              "type": "array",
              "items": {
                "type": "array",
                "minItems": 2,
                "maxItems": 2,
                "items": { "type": "number" }
              }
            }
          }
        }
      },
      "Ssr": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "type": { "type": "string" },
          "services": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Service" }
          },
          "geometry": { "$ref": "#/components/schemas/Polygon" },
          "altitude": { "type": "number" },
          "provider": { "type": "string" },
          "timestamp": { "type": "number" },
          "active": { "type": "boolean" }
        }
      },
      "SsrDto": {
        "type": "object",
        "required": ["type", "services", "geometry"],
        "properties": {
          "type": {
            "type": "string",
            "enum": ["ssr"]
          },
          "services": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Service" }
          },
          "geometry": { "$ref": "#/components/schemas/Polygon" },
          "altitude": { "type": "number" },
          "active": { "type": "boolean" }
        }
      }
    }
  }
}